% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_message.R
\name{display_message}
\alias{display_message}
\title{Display message.}
\usage{
display_message(a_message)
}
\arguments{
\item{a_message}{A character string containg the message to be displayed.}
}
\value{
Prints the message and returns the error_messsage character string
}
\description{
Display a message for the user and export it in the global environment.
}
\details{
The error_messsage variable that contains the message is exported in the global environment. This variable is read by the Java GUI.
}
\examples{
display_message("Hello world")
error_messsage


}
