% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_filter.R
\name{distance_filter}
\alias{distance_filter}
\title{Select barcode combinations with error correction properties}
\usage{
distance_filter(index_df, combinations_m, metric, d)
}
\arguments{
\item{index_df}{A dataframe containg barcode identifiers, corresponding DNA sequences along with GC content and presence of homopolymers.}

\item{combinations_m}{A matrix of compatible barcode combinations.}

\item{metric}{The type of distance (hamming or seqlev).}

\item{d}{The minimum value of the distance.}
}
\value{
Returns a filtered matrix containing the identifiers of compatible combinations of barcodes.
}
\description{
Filter barcode combinations for a given distance metric (hamming or seqlev) and threshold.
}
\details{
The Hamming distance is suitable for correcting substitution errors. 
The SeqLev distance is suitable for correcting insertion/deletion errors.
}
\examples{
write.table(DNABarcodeCompatibility::illumina,
 file <- tempfile(), row.names = FALSE, col.names = FALSE, quote=FALSE)
barcodes <- file_loading_and_checking(file)
m <- get_all_combinations(barcodes, 3, 4)
distance_filter(barcodes, m, "hamming", 3)

}
\references{
Buschmann, T. 2015. The Systematic Design and Application of Robust DNA Barcodes.

Buschmann, T. 2017. DNABarcodes: an R package for the systematic construction of DNA sample tags. Bioinformatics 33, 920–922.
}
\seealso{
\code{\link{get_all_combinations()}}, 
\code{\link{get_random_combinations()}}
}
