% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNABarcodeCompatibility-package.R
\docType{package}
\name{DNABarcodeCompatibility-package}
\alias{DNABarcodeCompatibility-package}
\title{DNABarcodeCompatibility: to find least-redundant sets of compatible barcodes for multiplex experiments performed on Illumina sequencing platforms}
\description{
The DNABarcodeCompatibility package provides six groups of utility functions to analyse and select DNA barcodes for deep sequencing purposes. It also provides a pipeline to perform a complete analysis and selection of DNA barcodes: experiment_design.
}
