% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_message.R
\docType{data}
\name{error_messsage}
\alias{error_messsage}
\title{Display message.}
\format{An object of class \code{character} of length 1.}
\usage{
display_message(a_message)
}
\arguments{
\item{a_message}{A character string containg the message to be displayed.}
}
\value{
Prints the message and returns the error_messsage character string
}
\description{
Display a message for the user and export it in the global environment.
}
\details{
The error_messsage variable that contains the message is exported in the global environment. This variable is read by the Java GUI.
}
\examples{
display_message("Hello world")
error_messsage


}
\keyword{datasets}
