% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_design.R
\name{experiment_design}
\alias{experiment_design}
\title{Get compatible combinations of DNA barcodes}
\usage{
experiment_design(file1, sample_number, multiplexing_level, file2 = NULL, export = NULL,  filter = NULL, metric = 3)
}
\arguments{
\item{sample_number}{number of libraries to be sequenced}

\item{multiplexing_level}{the number of samples to be multiplexed for sequencing}

\item{export}{TO BE DOCUMENTED}

\item{filter}{TO BE DOCUMENTED}

\item{metric}{minimal distance allowed between barcodes TO BE RENAMED dist}

\item{file1}{path to the file containing the barcode names and the corresponding DNA sequences.}

\item{file2}{path to the file containing the barcode names and the corresponding DNA sequences; used for dual-indexing.}
}
\value{
Returns a dataframe containing compatible DNA-barcode combinations organized by lanes of the flow cell.
}
\description{
Performs either an exhaustive or a random search of compatible DNA-barcode combinations depending on the size of the DNA-barcode population and of the number of samples to be multiplexed.
}
\details{
Illumina 4-channel sequencers use a green laser to read G/T nucleotides and a red laser to
read A/C nucleotides. With each sequencing cycle at least one of the two
nucleotides for each color channel must be read to ensure proper registration.
Since the DNA-barcode sequence of pooled, multiplexed libraries will be read
simultaneously, it is important to maintain color balance for each base of the
DNA-barcode sequences in the pooled library. Otherwise, DNA-barcode sequencing will fail due
to registration failure.

By specifying the total number of libraries and the number of libraries to be multiplexed, 
this function returns an optimal combination of DNA barcodes to be used for sequencing.
This function first converts DNA barcodes into binary words: A and C will be converted in "0", G and T in "1", see \code{\link{sequence_binary_conversion}}. 

Second, it selects a series of combinations of binarized barcodes that are compatible for 2-channel sequencing, see \code{\link{is_a_good_combination}}, \code{\link{get_combinations}}.

Third, it reduces redundancy between DNA barcodes so that consumables can be used as much homogeneously as possible in order to extend the life time of expensive library-preparation kits. To this end, the maximum entropy of the distribution of barcodes is calculated given the available number of barcodes and libraries, see \code{\link{entropia_result}}, \code{\link{entropia_max}}. The maximum-entropy value sets the upper limit to be reached or approached when calculating the entropy of concurrent sets of compatible DNA barcodes.  

Case 1: number of lanes < number of compatible DNA-barcode combinations
This function seeks for compatible DNA-barcode combinations of highest entropy. The search stops either if the maximum-entropy value is reached for a given set of DNA-barcode combinations or after 1000 iterations but returning the set of DNA-barcode combinations of the highest entropy.
The criterium of 1000 iterations has been choosen based on simulations.

Case 2: number of lanes > number of compatible DNA-barcode combinations
In such a case, there are not enough compatible DNA-barcode combinations and redundancy is inevitable. This may dramatically decrease the life time of the library-preparation kits leading to the necessity to buy new expensive kits just for providing enough of some DNA barcodes.

TO BE DOCUMENTED: 2-channel sequencers
}
\examples{
write.table(DNABarcodeCompatibility::illumina, file <- tempfile(), row.names = FALSE, col.names = FALSE, quote=FALSE)
results <- experiment_design(file, 6, 3)


}
\seealso{
\code{\link{get_result}}
}
