% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_loading_and_checking.R
\name{file_loading_and_checking}
\alias{file_loading_and_checking}
\title{Loading and checking DNA barcodes.}
\usage{
file_loading_and_checking(file)
}
\arguments{
\item{file}{The input data file that contains 2 columns separted by a space or a tabulation, namely the sequence identifiers and corresponding DNA sequence.}
}
\value{
Returns a dataframe containing sequence identifiers, nucleotide sequence, GC content, presence of homopolymers.
}
\description{
Load the file containing DNA barcodes and analyze barcode content.
}
\details{
This function loads the DNA barcodes from the input file and checks barcodes for unicity (identifier and sequence), DNA content, and equal size. It also calculates the GC content and detects the presence of homopolymers of length >= 3.
}
\examples{
write.table(DNABarcodeCompatibility::illumina,
 file <- tempfile(), row.names = FALSE, col.names = FALSE, quote=FALSE)
file_loading_and_checking(file)

}
