% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_random_combinations.R
\name{get_random_combinations}
\alias{get_random_combinations}
\title{Get a large set of compatible combinations.}
\usage{
get_random_combinations(index_df, mplex_level, chemistry)
}
\arguments{
\item{index_df}{A dataframe containing barcodes identifiers, corresponding DNA sequences along with GC content and presence of homopolymers.}

\item{mplex_level}{The number at which the barcodes will be multiplexed.}

\item{chemistry}{An integer representing the number of channels (1, 2, 4) of the desired Illumina plateform.}
}
\value{
Returns a matrix containing the identifiers of compatible barcode combinations.
}
\description{
Find a set of at most 1000 combinations of compatible barcodes by random pick.
}
\details{
This function is suited if the total number of possible combinations is too high for an exhaustive search in a reasonable amount of time.
}
\examples{
write.table(DNABarcodeCompatibility::IlluminaIndexes,
 txtfile <- tempfile(), row.names = FALSE, col.names = FALSE, quote=FALSE)
barcodes <- file_loading_and_checking(txtfile)
get_random_combinations(barcodes, 3, 4)


}
\seealso{
\code{\link{get_all_combinations}}
}
