% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IlluminaIndexes}
\alias{IlluminaIndexes}
\title{Barcode dataset from Illumina.}
\format{A data frame with 48 rows and 2 variables:
\describe{
  \item{V1}{barcode identifier}
  \item{V2}{DNA sequence}
  ...
}}
\usage{
IlluminaIndexes
}
\description{
48 barcodes from Illumina TruSeq Small RNA kits
}
\keyword{datasets}
