% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analyze_n_check_DNAbarcodes.R
\name{Utilities_DNAbarcodeAnalysis}
\alias{Utilities_DNAbarcodeAnalysis}
\alias{unicity_check}
\alias{character_check}
\alias{sequences_character_check}
\alias{length_check}
\alias{character_and_length_check}
\alias{index_check}
\alias{get_sequence_GC_content}
\alias{get_index_GC_content}
\alias{get_sequence_homopolymer}
\alias{get_index_homopolymer}
\alias{index_distance}
\alias{low_hamming_distance}
\alias{low_seqlev_distance}
\title{Group of utility functions}
\usage{
unicity_check(index)

character_check(sequence)

sequences_character_check(sequences)

length_check(sequences)

character_and_length_check(index)

index_check(index)

get_sequence_GC_content(sequence)

get_index_GC_content(index)

get_sequence_homopolymer(sequence)

get_index_homopolymer(index)

index_distance(index)

low_hamming_distance(index_distances, metric)

low_seqlev_distance(index_distances, metric)
}
\arguments{
\item{index}{a dataframe containing a representation of the DNA-barcode dataset.}

\item{sequence}{a DNA sequence as string}

\item{index_distances}{a dataframe containing the calculted distances between pairs de DNA barcodes}

\item{metrict}{threshold distance between barcodes}
}
\value{
The return value of each function is briefly documented thereafter for each function
}
\description{
Group of functions for analyzing and checking DNA barcodes
}
\details{
These functions allow the user to check the DNA content of the barcodes, and in particular the GC content, the homopolymer length and the distance between barcodes.
}
\section{Functions}{

\itemize{
\item unicity_check(): check the unicity of barcodes sequences and identifiers; Returns a boolean
}


\itemize{
\item character_check(): check the character composition of one sequence; Returns a boolean
}


\itemize{
\item sequences_character_check(): check the character composition of each sequence in a vector of sequences;
}


\itemize{
\item length_check(): check the length of each sequence; returns a boolean
}


\itemize{
\item character_and_length_check(): check for character type and sequence length; returns a boolean
}


\itemize{
\item index_check(): check the index for possible issues; returns a boolean
}


\itemize{
\item get_sequence_GC_content(): calculate GC content of one sequence; returns GC content as numerics
}


\itemize{
\item get_index_GC_content(): calculate GC content of each sequence and assign the sequence identifier; returns TO BE COMPLETED
}


\itemize{
\item get_sequence_homopolymer(): calculate homopolymer length of one sequence; returns a boolean
}


\itemize{
\item get_index_homopolymer(): calculate homopolymer length of each sequence and assign the sequence identifier, returns
}


\itemize{
\item index_distance(): generates all possible barcode pairs and caculates hamming and seqlev distances; returns a dataframe
}


\itemize{
\item low_hamming_distance():  TO BE DOCUMENTED
}


\itemize{
\item low_hamming_distance():  TO BE DOCUMENTED
}
}

