% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_barcode_combinations.R
\name{Utilities_DNAbarcodeCombinations}
\alias{Utilities_DNAbarcodeCombinations}
\alias{binary_word_into_numeric}
\alias{vectors_into_matrix}
\alias{any_different}
\alias{is_a_good_combination}
\alias{list_of_good_combinations}
\alias{id_into_binary_sequence}
\alias{matrix_id_to_binary_sequence}
\alias{filter_combinations}
\title{Group of utility functions}
\usage{
binary_word_into_numeric(binary_word)

vectors_into_matrix(binary_word)

any_different(binary_sequence)

is_a_good_combination(combination_matrix)

list_of_good_combinations(matrix_id)

id_into_binary_sequence(index_id_combination, index)

matrix_id_to_binary_sequence(matrix_id, index)

filter_combinations(combinations, low_distance_tab)
}
\arguments{
\item{index}{a dataframe containing a representation of the DNA-barcode dataset.}

\item{sequence}{the name of the file which the data are to be read from. Each row of the table appears as one line of the file. If it does not contain an absolute path, the file name is relative to the current working directory, \code{\link[base]{getwd}}.}
}
\value{
A data frame containing a representation of the loaded data.
}
\description{
Group of functions for loading and checking DNA barcodes
}
\details{
These functions allow the user to read the input file and check its content
}
\section{Functions}{

\itemize{
\item binary_word_into_numeric(): conversion of the string sequence into a vector of numeric
}


\itemize{
\item vectors_into_matrix(): conversion of each index sequence combination into a matrix
}


\itemize{
\item any_different(): test if a column/line of a index combination is correct
}


\itemize{
\item is_a_good_combination(): check if a combination of indexes is correct
}


\itemize{
\item list_of_good_combinations(): keep only good combinations
}


\itemize{
\item id_into_binary_sequence(): matches an id to its binary_sequence
}


\itemize{
\item matrix_id_to_binary_sequence(): matches the matrix_id to the matrix_binary_sequence
}


\itemize{
\item filter_combinations(): low distance tab = hamming rejection table or seq lev rejection table
}
}

\examples{








}
