% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_binary_conversion.R
\name{Utilities_DNAsequenceBinaryConversion}
\alias{Utilities_DNAsequenceBinaryConversion}
\alias{sequence_binary_conversion}
\alias{index_binary_conversion}
\title{Group of utility functions}
\usage{
sequence_binary_conversion(sequence)

index_binary_conversion(index)
}
\arguments{
\item{sequence}{the name of the file which the data are to be read from. Each row of the table appears as one line of the file. If it does not contain an absolute path, the file name is relative to the current working directory, \code{\link[base]{getwd}}.}

\item{index}{a dataframe containing a representation of the DNA-barcode dataset.}
}
\value{
The return value of each function is briefly documented thereafter for each function.
}
\description{
Group of functions for loading and checking DNA barcodes
}
\details{
These functions allow the user to read the input file and check its content
}
\section{Functions}{

\itemize{
\item sequence_binary_conversion(): converts a DNA sequence into a binary sequence for which A or C becomes 0 and G or T becomes 1
}


\itemize{
\item index_binary_conversion(): binarize each DNA sequence in a vector
}
}

