% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_entropy.R
\name{Utilities_Entropy}
\alias{Utilities_Entropy}
\alias{shannon_entropy}
\alias{entropy_result}
\alias{entropy_max}
\alias{comb_result}
\alias{get_result}
\title{Group of utility functions}
\usage{
shannon_entropy(frequence)

entropy_result(index_combination)

entropy_max(index_number, sample_number)

comb_result(index_combination, sample_number, index_number, nb_lane)

get_result(index, sample_number, multiplexing_level, filter = NULL,
  metric = 3)
}
\arguments{
\item{frequence}{probability of occurrence of a given barcode in a set of compatible barcode combinations.}

\item{index_combination}{matrix of compatible DNA-barcode combinations.}

\item{index_number}{total number of DNA barcodes provided by the user.}

\item{sample_number}{number of libraries to be sequenced.}

\item{nb_lane}{an integer, given the sample_number and multiplex level.}
}
\value{
The return value of each function is briefly documented thereafter for each function.
}
\description{
Group of functions for Shannon's entropy calculations
}
\details{
These functions allow the user to calculate the entropy and max entropy of barcode combinations


In the context of DNA barcodes, the maximum entropy is reached when the selected barcodes are the least redundant. It corresponds to an ideal case for which consumables are used as much homogeneously as possible in order to extend the life time of expensive library-preparation kits.

Case 1: index_number > sample_number
In such a case, it would be ideal to assign distinct barcodes to the various libraries to be sequenced (regardless multiplexing), and the resulting Shannon entropy would correspond to the maximum entropy (all barcodes are different): \code{Hmax = log(sample_number)}.

Case 2: index_number < sample_number
Since there are not as many distinct barcodes as libraries to be sequenced, it is inevitable to use the same barcodes for different libraries. The maximum entropy expression is non-trivial and it depends on both the available number of barcodes and the number of libraries.
}
\section{Functions}{

\itemize{
\item shannon_entropy(): Shannon's entropy
}


\itemize{
\item entropy_result(): Shannon's entropy for a matrix of combinations
}


\itemize{
\item entropy_max(): calculte the max entropy given the number of barcodes and libraries
}


\itemize{
\item comb_result(): Integrate the compatible barcodes combinations with user-deduced experimental constaints: number of libraries to be sequenced, the total number of DNA barcodes provided by the user and the number of lanes in the experiment.
}


\itemize{
\item get_result(): get a compatible set of DNA barcodes for 4-channel sequencing plateforms. DNA-barcode redundancy is reduced by using a Shannon-entropy based optimization.
}
}

\examples{
# For a set of 48 barcodes and 96 samples (libraries)
entropy_max(index_number = 48, sample_number = 96)

}
