% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_input.R
\name{Utilities_Input}
\alias{Utilities_Input}
\alias{file_loading_and_checking}
\alias{read_index}
\alias{display_message}
\title{Group of utility functions}
\usage{
file_loading_and_checking(file)

read_index(file)

display_message(a_message)
}
\arguments{
\item{file}{the name of the file which the data are to be read from. Each row of the table appears as one line of the file. If it does not contain an absolute path, the file name is relative to the current working directory, \code{\link[base]{getwd}}.}
}
\value{
A data frame containing a representation of the loaded data.
}
\description{
Group of functions for loading and checking DNA barcodes
}
\details{
These functions allow the user to read the input file and check its content
}
\section{Functions}{

\itemize{
\item file_loading_and_checking(): Load and check DNA barcodes
}


\itemize{
\item read_index(): load DNA barcodes only
}


\itemize{
\item display_message(): display a message on the screen
}
}

\examples{
write.table(DNABarcodeCompatibility::illumina, file <- tempfile(), row.names = FALSE, col.names = FALSE, quote=FALSE)
index <- file_loading_and_checking(file)
index

}
