% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sample_and_multiplexing_level_check.R
\name{Utilities_Multiplexing}
\alias{Utilities_Multiplexing}
\alias{sample_and_multiplexing_level_check}
\alias{sample_number_check}
\alias{multiplexing_level_set}
\title{Group of utility functions}
\usage{
sample_and_multiplexing_level_check(sample_number, multiplexing_level)

sample_number_check(sample_number)

multiplexing_level_set(sample_number)
}
\arguments{
\item{sample_number}{the number of libraries to be sequenced.}

\item{multiplexing_level}{the number librairies to be pooled for sequencing}
}
\value{
The return value of each function is briefly documented thereafter for each function
}
\description{
Group of functions for finding the possible multiplex levels given the number of libraries
}
\details{
These functions help the user to identify the possible values of multiplex levels given the number of libraries.
}
\section{Functions}{

\itemize{
\item sample_and_multiplexing_level_check(): check the adequacy between the number of libraries and the multiplexing level; Returns a boolean
}


\itemize{
\item sample_number_check(): check if the number of libraries is realistic; Returns a boolean
}


\itemize{
\item multiplexing_level_set(): calculate the multiplex level choices given the number of libraires; Returns a vector of multiplex levels
}
}

