% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNABarcodeCompatibility-package.R
\docType{package}
\name{DNABarcodeCompatibility-package}
\alias{DNABarcodeCompatibility-package}
\title{DNABarcodeCompatibility: to find least-redundant sets of compatible barcodes for multiplex experiments performed on Illumina sequencing platforms.}
\description{
The DNABarcodeCompatibility package provides six functions to load DNA barcodes, and to generate, filter and optimize sets of compatible barcode combinations for multiplex sequecing experiments.
In particular, barcode combinations can be filtered to keep those that are robust against substitution and insertion/deletion errors thereby facilitating the demultiplexing step.
In addtion, the package provides an optimizer function to further favor the selection of compatible barcode combinations with least redundancy of DNA barcodes.
}
