% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_combinations.R
\name{get_all_combinations}
\alias{get_all_combinations}
\title{Get all compatible combinations.}
\usage{
get_all_combinations(index_df, mplex_level, chemistry)
}
\arguments{
\item{index_df}{A dataframe containg barcodes identifiers, corresponding DNA sequences along with GC content and presence of homopolymers.}

\item{mplex_level}{The number at which the barcodes will be multiplexed.}

\item{chemistry}{An integer representing the number of channels (1, 2, 4) of the desired Illumina plateform.}
}
\value{
Returns a matrix containing the identifiers of compatible barcode combinations.
}
\description{
Find the exhaustive set of compatible barcode combinations.
}
\details{
Be aware that the number of combinations may tend to infinity.
}
\examples{
write.table(DNABarcodeCompatibility::illumina,
 txtfile <- tempfile(), row.names = FALSE, col.names = FALSE, quote=FALSE)
barcodes <- file_loading_and_checking(txtfile)
get_all_combinations(barcodes, 3, 4)


}
\seealso{
\code{\link{get_random_combinations}}
}
