% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IlluminaIndexes}
\alias{IlluminaIndexes}
\title{Barcode dataset from Illumina with features.}
\format{A data frame with 48 rows and 4 variables:
\describe{
\item{Id}{barcode identifier}
\item{sequence}{DNA sequence}
\item{GC_content}{percentage of G and C relative to A and T}
\item{homopolymer}{presence of nucleotide repetitions (>= 3)}
...
}}
\usage{
IlluminaIndexes
}
\description{
48 barcodes from Illumina TruSeq Small RNA kits along 
with percentage in CG content and presence of homopolymers of size >= 3
}
\keyword{datasets}
