% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_combinations.R
\name{optimize_combinations}
\alias{optimize_combinations}
\title{Find a set of barcode combinations with least redundant barcodes}
\usage{
optimize_combinations(combination_m, nb_lane, index_number)
}
\arguments{
\item{combination_m}{A matrix of compatible barcode combinations.}

\item{nb_lane}{The number of lanes to be use for sequencing (i.e. the number of libraries divided by the multiplex level).}

\item{index_number}{The total number of distinct DNA barcodes in the dataset.}
}
\value{
Returns a matrix containing an optimized set of combinations of compatible barcodes.
}
\description{
This function uses the Shannon Entropy to identify a set of compatible barcode combinations with least redundancy between DNA barcodes.
}
\details{

}
\examples{
write.table(DNABarcodeCompatibility::illumina,
 txtfile <- tempfile(), row.names = FALSE, col.names = FALSE, quote=FALSE)
barcodes <- file_loading_and_checking(txtfile)
m <- get_random_combinations(barcodes, 3, 4)
optimize_combinations(m, 12, 48)


}
\seealso{
\code{\link{get_all_combinations}}, 
\code{\link{get_random_combinations}}, 
\code{\link{experiment_design}}
}
